refactoring publish and inquiry tools utilities functions
