+++ agent-pkg/src/commands/publish.ts
@@ -1,9 +1,9 @@
 import chalk from 'chalk';
 import ora from 'ora';
 import { CommandResult } from '../types/index.js';
-import { updateRegistry } from '../utils/registry.js';
+import { updateRegistry } from '../utils/commands/registry/updateRegistry.js';
 import { validateToolStructure } from '../utils/validation.js';
-import { packageAndUploadTool } from '../utils/commands/publish/toolPublish.js';
+import { packageAndUploadTool } from '../utils/commands/registry/toolPublish.js';
 
 export const publish = async (toolPath: string, options: any = {}): Promise<CommandResult> => {
   const spinner = ora(`Publishing tool from ${toolPath}...`).start();
