+++ agent-pkg/src/utils/commands/registry/toolInquiry.ts
@@ -1,17 +1,14 @@
 import fs from 'fs/promises';
-import { ToolMetadata, ToolRegistry } from '../types/index.js';
+import { ToolMetadata, ToolRegistry } from '../../../types/index.js';
 import {
   getToolInfo,
   getToolVersion,
   getToolVersions,
   getLatestToolVersion,
   getAllToolNames,
-  registerTool,
-  addToolVersion,
-  isToolOwner,
-} from './blockchain/contractClient.js';
+} from '../../blockchain/contractClient.js';
 import chalk from 'chalk';
-import { REGISTRY_CACHE_PATH } from './shared/path.js';
+import { REGISTRY_CACHE_PATH } from '../../shared/path.js';
 
 const fetchRegistryFromBlockchain = async (): Promise<ToolRegistry> => {
   try {
@@ -65,7 +62,7 @@ const fetchRegistryFromBlockchain = async (): Promise<ToolRegistry> => {
   }
 };
 
-const getLocalRegistryCache = async (): Promise<ToolRegistry> => {
+export const getLocalRegistryCache = async (): Promise<ToolRegistry> => {
 
   try {
     const cacheData = await fs.readFile(REGISTRY_CACHE_PATH, 'utf8');
@@ -200,70 +197,5 @@ export const getToolVersionFromRegistry = async (
     }
 
     return null;
-  }
-};
-
-
-/**
- * Update the registry on the blockchain
- * @param toolMetadata Tool metadata to update
- * @returns Transaction hash of the update
- */
-export const updateRegistry = async (toolMetadata: ToolMetadata): Promise<string> => {
-  try {
-    console.log(chalk.blue('Updating registry on blockchain...'));
-
-    let txHash = '';
-
-    const metadata = JSON.stringify({
-      description: toolMetadata.description || '',
-      author: toolMetadata.author || '',
-      updated: toolMetadata.updated || new Date().toISOString(),
-    });
-
-    try {
-      const exists = await getToolInfo(toolMetadata.name);
-
-      if (exists.latestVersion && (await isToolOwner(toolMetadata.name))) {
-        console.log(
-          chalk.yellow(
-            `Tool ${toolMetadata.name} already exists. Adding new version ${toolMetadata.version}`,
-          ),
-        );
-        txHash = await addToolVersion(
-          toolMetadata.name,
-          toolMetadata.cid,
-          toolMetadata.version,
-          metadata,
-        );
-      } else if (exists.latestVersion) {
-        throw new Error(`Tool ${toolMetadata.name} already exists and you're not the owner`);
-      } else {
-        txHash = await registerTool(
-          toolMetadata.name,
-          toolMetadata.cid,
-          toolMetadata.version,
-          metadata,
-        );
-      }
-    } catch (error) {
-      console.log(chalk.green(`Registering new tool ${toolMetadata.name}`));
-      txHash = await registerTool(
-        toolMetadata.name,
-        toolMetadata.cid,
-        toolMetadata.version,
-        metadata,
-      );
-    }
-
-    const registry = await getLocalRegistryCache();
-    registry.tools[toolMetadata.name] = toolMetadata;
-    registry.updated = new Date().toISOString();
-    await fs.writeFile(REGISTRY_CACHE_PATH, JSON.stringify(registry, null, 2));
-
-    return txHash;
-  } catch (error) {
-    console.error('Error updating registry:', error);
-    throw error;
   }
 };
