+++ agent-pkg/src/utils/commands/registry/toolPublish.ts
@@ -95,3 +95,4 @@ export const packageAndUploadTool = async (
 
   return { cid, metadata };
 };
+
